+++ edn/src/lib.rs
@@ -30,6 +30,12 @@ pub mod types;
 pub mod pretty_print;
 pub mod utils;
 pub mod matcher;
+pub mod value_rc;
+pub use value_rc::{
+    Cloned,
+    FromRc,
+    ValueRc,
+};
 
 pub mod parse {
     include!(concat!(env!("OUT_DIR"), "/edn.rs"));
